% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pestcontrolGame.R
\name{pestcontrolGame}
\alias{pestcontrolGame}
\title{Tabulate results for a simple in-class stag hunt game.}
\usage{
pestcontrolGame(
  resultsSheet,
  partners = "random",
  roleSheet = NULL,
  roleLabs = c("Anil", "Bala"),
  payoff = c(3, 4, 1, 2),
  seed = 8675309,
  auth = FALSE,
  names = NULL,
  ...
)
}
\arguments{
\item{payoff}{is a vector indicating the interdependent payoffs from the different pairs of {Student, Partner} strategies: \code{c({Compete, Compete}, {Collude, Compete}, {Compete, Collude}, {Collude, Collude})}}

\item{sheet}{(required) is a character string sheet ID corresponding to the Google Sheets location containing the individual submissions.}

\item{outfile}{is a character string giving a name to the new Google Sheet where the instructor wants to store the scores.}
}
\value{
\code{type} returns the type of activity (pestcontrolGame).

\code{results} returns the original submissions (with equilibria and points per round added).

\code{grades} returns the aggregated points "won" by each student for the entire activity.
}
\description{
Tabulate results for a simple in-class stag hunt game.
}
\details{
\code{staghuntGame} tabulates the results of a simple stag hunt game in which students' points depend on their strategy and the strategy chosen by their randomly-assigned partners.
}
