% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicgoodGame.R
\name{publicgoodGame}
\alias{publicgoodGame}
\title{Tabulate public good game results.}
\usage{
publicgoodGame(
  sheet,
  endowment = 0,
  return = 1.2,
  auth = FALSE,
  names = NULL,
  email = NULL,
  ...
)
}
\arguments{
\item{sheet}{(required) is a character string url corresponding to the Google Sheets location containing the individual submissions.}

\item{endowment}{is the number of gifted points students start the game with (default is 5).}

\item{return}{is the scaling factor for points contributed to the public good (default is 1.2).}

\item{auth}{is a logical indicating whether to use an authentication token to access the Sheet containing the individual submissions.}

\item{names}{character list of the column names in \code{sheet}.}

\item{email}{is an email address that matches the user account containing the Sheet with the individual submissions (if \code{auth == TRUE}).}
}
\value{
\code{type} returns the type of activity (publicgoodGame).

\code{results} returns the original submissions.

\code{blinded esults} returns the submissions without student names.

\code{grades} returns the aggregated, stacked points "won" by each student for the entire activity.
}
\description{
Tabulates and assigns points for the results of a simple in-class public good game.
}
\details{
\code{publicgoodGame} tabulates the results of a simple public good game based on Leuthold (1993) and Holt and Laury (1997). Each student starts with an endowment of 5 points that they can either keep or anonymously and voluntarily contribute to a public good. For each point a student keeps, they receive only that point. For each point contributed to the public good, \emph{everyone} receives R/N points, where N is the number of participants and R > 1 is the return the group earns on public good contributions. In other words, students have the opportunity to boost their points if everyone contributes all of them. But, if only a few participants contribute, their net return could diminish.
}
\references{
Holt, Charles A. and Laury, Susan K. (1997). Classroom Games: Voluntary Provision of a Public Good.
\emph{Journal of Economic Perspectives} 11(4), pp. 209-215.
}
