% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomLeaders.R
\name{randomLeaders}
\alias{randomLeaders}
\title{Randomly assigns partner roles}
\usage{
randomLeaders(
  sheet,
  size = 2,
  seed = 8675309,
  round = 1,
  auth = FALSE,
  names = NULL
)
}
\arguments{
\item{sheet}{(required) is an object containing the list of student names with surnames and given names (see options for \code{last} and \code{first} below).}

\item{size}{is the size of each group (default is 2).}

\item{seed}{is a random seed (default is 8675309).}

\item{last}{is the name of the column in \code{sheet} containing the surnames of the participants (default is "Last.Name"). If the column names in \code{sheet} contain spaces, \code{randomGroups} automatically replaces spaces with periods using \code{make.names()} so that "Last Name" becomes "Last.Name".}

\item{first}{is the name of the column in \code{sheet} containing the given names of the participants (default is "First.Name").}
}
\value{
\code{studentList} returns list of names and group numbers.

\code{results} returns the original submissions (with equilibria and points per round added).

\code{grades} returns the aggregated points "won" by each student for the entire activity.
}
\description{
Randomly assigns partner roles
}
\details{
\code{randomRoles} uses a random number to assign partner roles (e.g. leader-follower, proposer-responder) to a list of student names a Google Sheet containing a student roster into
}
