% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomRoles.R
\name{randomRoles}
\alias{randomRoles}
\title{Randomly assigns partner roles}
\usage{
randomRoles(
  sheet,
  size = 2,
  seed = 8675309,
  roleLabs = c("Leader", "Follower"),
  auth = FALSE,
  names = NULL,
  email = NULL,
  ...
)
}
\arguments{
\item{sheet}{(required) is an object containing the list of students.}

\item{size}{is the size of each group (default is \code{2}).}

\item{seed}{is a random seed (default is \code{8675309}).}

\item{auth}{is a logical indicating whether to use an authentication token to access the Sheet containing the individual submissions.}

\item{names}{character list of the column names in \code{sheet}.}

\item{email}{is an email address that matches the user account containing the Sheet with the individual submissions (if \code{auth == TRUE}).}

\item{roles}{contains the labels for the roles (default is \code{c("Leader", "Follower")})}
}
\value{
\code{studentList} returns list of names and group numbers.

\code{results} returns the original submissions (with equilibria and points per round added).

\code{grades} returns the aggregated points "won" by each student for the entire activity.
}
\description{
Randomly assigns partner roles
}
\details{
\code{randomRoles} uses a random number to assign partner roles (e.g. leader-follower, proposer-responder) to a list of student names a Google Sheet containing a student roster into
}
