% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMarketGameApp.R
\name{runMarketGameApp}
\alias{runMarketGameApp}
\title{Run an \code{econGame} app.}
\usage{
runMarketGameApp(game = NULL)
}
\arguments{
\item{game}{the name of the \code{econGame} app you want to tabulate, which can be one of:
\enumerate{
\item \code{equlibriumGame} to tabulate results simple supply-and-demand equilibrium game.
\item \code{entryGame} two-sector entry-and-exit game.
\item \code{bertrandGame} Bertrand duopoly game.
\item \code{cournotGame} Cournot duopoly game.
\item \code{stackelbergGame} sequential leader-follower game.
\item \code{ultimatumGame} ultimatum game.
\item \code{anchoringGame} the effects of anchoring on guessing.
\item \code{lobbyGame} lottery-auction game.
\item \code{publicgoodGame} public good game with free-rider effects.
\item \code{pollutionGame} the effects of alternative pollution-abatement policies.
\item \code{multipdGame} multi-player prisoner's dilemma game.
\item \code{staghuntGame} two-person stag hunt gmae.
}}
}
\description{
This is a wrapper function for the game-specific \code{run_Game_App()} functions in the marketGames package.
}
